refactor(zero-throughput): separate binary and linear sweeps, drop multi-RM, and clean up flags
