+++ apps/zero-throughput/src/config.ts
@@ -41,8 +41,6 @@ const options = {
   topology: v.literalUnion('single', 'distributed').default('single'),
   numViewSyncers: v.number().default(1),
   numSyncWorkers: v.number().optional(),
-  highAvailabilityRM: v.boolean().default(false),
-  numReplicationManagers: v.literalUnion(1, 2).optional(),
   profileRM: v.boolean().default(false),
   profileVS: v.boolean().default(false),
 
@@ -78,8 +76,6 @@ export type BenchmarkConfig = {
   readonly topology: BenchmarkTopology;
   readonly numViewSyncers: number;
   readonly numSyncWorkers: number;
-  readonly highAvailabilityRM: boolean;
-  readonly numReplicationManagers: 1 | 2;
   readonly users: number;
   readonly queriesPerUser: number;
   readonly rowsPerQuery: number;
@@ -139,9 +135,6 @@ export function loadConfig(): BenchmarkConfig {
   assertPositiveInteger('numViewSyncers', parsed.numViewSyncers);
   const numSyncWorkers = parsed.numSyncWorkers ?? parsed.zero.numSyncWorkers;
   assertPositiveInteger('numSyncWorkers', numSyncWorkers);
-  const highAvailabilityRM =
-    parsed.highAvailabilityRM || parsed.numReplicationManagers === 2;
-  const numReplicationManagers: 1 | 2 = highAvailabilityRM ? 2 : 1;
   assertNonNegativeInteger('payloadBytes', parsed.payloadBytes);
   assertPositiveInteger('durationMs', parsed.durationMs);
   assertNonNegativeInteger('warmupMs', parsed.warmupMs);
@@ -168,8 +161,6 @@ export function loadConfig(): BenchmarkConfig {
     topology: parsed.topology,
     numViewSyncers: parsed.numViewSyncers,
     numSyncWorkers,
-    highAvailabilityRM,
-    numReplicationManagers,
     users: parsed.users,
     queriesPerUser: parsed.queriesPerUser,
     rowsPerQuery: parsed.rowsPerQuery,
