+++ apps/zero-throughput/src/sweep.ts
@@ -1,1316 +1,67 @@
-import {execFileSync, spawn} from 'node:child_process';
-import {once} from 'node:events';
-import {createWriteStream, type WriteStream} from 'node:fs';
-import {access, appendFile, mkdir, readFile, writeFile} from 'node:fs/promises';
-import {dirname, join} from 'node:path';
-import {fileURLToPath} from 'node:url';
-import type {
-  BenchmarkModel,
-  BenchmarkProfile,
-  BenchmarkTopology,
-} from './config.ts';
-import {appPath, appRoot, repoRoot} from './config.ts';
-import {startPostgres, stopPostgres} from './processes.ts';
-import type {BenchmarkResult} from './results.ts';
-import {formatDuration} from './util.ts';
-
-const DEFAULT_PROFILES = ['relational', 'email', 'forum'] as const;
-const DEFAULT_MODELS = ['hot'] as const;
-const DEFAULT_USERS = [50, 100, 200, 400] as const;
-const DEFAULT_ROWS_PER_QUERY = [50] as const;
-const DEFAULT_SYNC_WORKERS = [1, 2, 4] as const;
-const DEFAULT_WRITE_RATE_MAX = 100;
-const DEFAULT_SEARCH_STEPS = 7;
-const CSV_ESCAPE_PATTERN = /[",\n]/;
-const PROFILE_VALUES = new Set<BenchmarkProfile>([
-  'feed-append',
-  'email',
-  'forum',
-  'relational',
-]);
-const MODEL_VALUES = new Set<BenchmarkModel>(['hot', 'realistic']);
-
-type SweepConfig = {
-  readonly runID: string;
-  readonly mode: 'binary' | 'linear';
-  readonly writeRates?: readonly number[] | undefined;
-  readonly topology?: BenchmarkTopology | undefined;
-  readonly numViewSyncers?: number | undefined;
-  readonly numSyncWorkers?: number | undefined;
-  readonly highAvailabilityRM?: boolean | undefined;
-  readonly numReplicationManagers?: (1 | 2) | undefined;
-  readonly writeConcurrency?: number | undefined;
-  readonly profileRM?: boolean | undefined;
-  readonly profileVS?: boolean | undefined;
-  readonly profiles: readonly BenchmarkProfile[];
-  readonly models: readonly BenchmarkModel[];
-  readonly users: readonly number[];
-  readonly rowsPerQuery: readonly number[];
-  readonly syncWorkers: readonly number[];
-  readonly queriesPerUser: number;
-  readonly durationMs: number;
-  readonly warmupMs: number;
-  readonly settleMs: number;
-  readonly sampleIntervalMs: number;
-  readonly progressIntervalMs: number;
-  readonly sloP99LagMs: number;
-  readonly batchSize: number;
-  readonly payloadBytes: number;
-  readonly writeRateMin: number;
-  readonly writeRateMax: number;
-  readonly searchSteps: number;
-  readonly repetitions: number;
-  readonly outputDir: string;
-  readonly zeroPort: number;
-  readonly pgStart: boolean;
-  readonly pgURL: string | undefined;
-  readonly resume: boolean;
-  readonly continueOnError: boolean;
-  readonly dryRun: boolean;
-  readonly limit: number | undefined;
-  readonly verboseChildLogs: boolean;
-};
-
-type SweepPoint = {
-  readonly profile: BenchmarkProfile;
-  readonly model: BenchmarkModel;
-  readonly users: number;
-  readonly queriesPerUser: number;
-  readonly rowsPerQuery: number;
-  readonly zeroNumSyncWorkers: number;
-};
-
-type AttemptStatus = 'pass' | 'fail' | 'error';
-
-type SweepAttempt = {
-  readonly point: SweepPoint;
-  readonly writeRate: number;
-  readonly repetition: number;
-  readonly outputPath: string;
-  readonly logPath: string;
-  readonly reused: boolean;
-  readonly status: AttemptStatus;
-  readonly exitCode: number | undefined;
-  readonly error: string | undefined;
-  readonly summary: BenchmarkResult['summary'] | undefined;
-};
-
-type PointResult = {
-  readonly point: SweepPoint;
-  readonly bestWriteRate: number | undefined;
-  readonly lowerBoundWriteRate: number;
-  readonly upperBoundWriteRate: number;
-  readonly attempts: readonly SweepAttempt[];
-  readonly bestAttempt: SweepAttempt | undefined;
-};
-
+/**
+ * Sweep entrypoint that delegates to either binary capacity search
+ * (binary-sweep.ts) or linear matrix comparison (linear-sweep.ts).
+ */
 const argv = process.argv.slice(2);
-const config = parseArgs(argv[0] === '--' ? argv.slice(1) : argv);
+const cleanArgv = argv[0] === '--' ? argv.slice(1) : argv;
 
-if (config.dryRun) {
-  printDryRun(config);
-} else if (config.mode === 'linear') {
-  await runLinearSweep(config);
-} else {
-  await runSweep(config);
-}
+const isLinear =
+  cleanArgv.includes('--mode=linear') ||
+  cleanArgv.some((arg, i) => arg === '--mode' && cleanArgv[i + 1] === 'linear');
 
-async function runSweep(config: SweepConfig): Promise<void> {
-  const outputDir = appPath(config.outputDir);
-  const runsDir = join(outputDir, 'runs');
-  const childLogsDir = join(outputDir, 'child-logs');
-  const attemptsPath = join(outputDir, 'attempts.jsonl');
-  const pointsPath = join(outputDir, 'points.jsonl');
-  const summaryPath = join(outputDir, 'summary.csv');
-  const manifestPath = join(outputDir, 'manifest.json');
-  const points = sweepPoints(config);
-  const results: PointResult[] = [];
-  let postgresStarted = false;
-
-  await mkdir(runsDir, {recursive: true});
-  await mkdir(childLogsDir, {recursive: true});
-  await writeFile(
-    manifestPath,
-    `${JSON.stringify(
-      {
-        runID: config.runID,
-        startedAt: new Date().toISOString(),
-        gitCommit: gitCommit(),
-        config,
-        points,
-        maxBenchmarkRuns:
-          points.length * config.searchSteps * config.repetitions,
-      },
-      null,
-      2,
-    )}\n`,
-  );
-  await writeFile(attemptsPath, '');
-  await writeFile(pointsPath, '');
-  await writeFile(summaryPath, `${csvHeader()}\n`);
-
-  const onSigint = () => {
-    stderr('\nInterrupted. Cleaning up sweep PostgreSQL if needed...\n');
-    void (async () => {
-      if (postgresStarted) {
-        await stopPostgres();
-      }
-      process.exit(130);
-    })();
-  };
-  process.once('SIGINT', onSigint);
-
-  try {
-    stdout(`zero-throughput sweep ${config.runID}\n`);
-    stdout(`output: ${outputDir}\n`);
-    stdout(
-      `points: ${points.length}, max benchmark runs: ${
-        points.length * config.searchSteps * config.repetitions
-      }\n`,
-    );
-
-    if (config.pgStart) {
-      stdout('Starting sweep PostgreSQL...\n');
-      await startPostgres();
-      postgresStarted = true;
-    }
-
-    for (const [index, point] of points.entries()) {
-      stdout(
-        `\n[${index + 1}/${points.length}] ${pointLabel(point)} ` +
-          `searching ${config.writeRateMin}-${config.writeRateMax} writes/s\n`,
-      );
-      const result = await runPointSearch({
-        config,
-        point,
-        runsDir,
-        childLogsDir,
-        attemptsPath,
-      });
-      results.push(result);
-      await appendFile(pointsPath, `${JSON.stringify(result)}\n`);
-      await appendFile(summaryPath, `${csvRow(result)}\n`);
-      stdout(
-        `best sustainable write rate: ${
-          result.bestWriteRate === undefined ? 'none' : result.bestWriteRate
-        } logical writes/s\n`,
-      );
-    }
-
-    await writeFile(
-      join(outputDir, 'sweep.json'),
-      `${JSON.stringify(
-        {
-          runID: config.runID,
-          finishedAt: new Date().toISOString(),
-          config,
-          results,
-        },
-        null,
-        2,
-      )}\n`,
-    );
-    stdout(`\nsummary: ${summaryPath}\n`);
-    stdout(`attempts: ${attemptsPath}\n`);
-  } finally {
-    process.off('SIGINT', onSigint);
-    if (postgresStarted) {
-      stdout('Stopping sweep PostgreSQL...\n');
-      await stopPostgres();
-    }
-  }
-}
-
-async function runLinearSweep(config: SweepConfig): Promise<void> {
-  const outputDir = appPath(config.outputDir);
-  const runsDir = join(outputDir, 'runs');
-  const childLogsDir = join(outputDir, 'child-logs');
-  const attemptsPath = join(outputDir, 'attempts.jsonl');
-  const points = sweepPoints(config);
-  const writeRates =
-    config.writeRates && config.writeRates.length > 0
-      ? config.writeRates
-      : [100, 250, 500, 1000, 2000];
-
-  await mkdir(runsDir, {recursive: true});
-  await mkdir(childLogsDir, {recursive: true});
-  await writeFile(attemptsPath, '');
-
-  let postgresStarted = false;
-  const onSigint = () => {
-    stderr('\nInterrupted. Cleaning up sweep PostgreSQL if needed...\n');
-    void (async () => {
-      if (postgresStarted) {
-        await stopPostgres();
-      }
-      process.exit(130);
-    })();
-  };
-  process.once('SIGINT', onSigint);
-
-  const attempts: SweepAttempt[] = [];
-
-  try {
-    stdout(`zero-throughput linear sweep ${config.runID}\n`);
-    stdout(`output: ${outputDir}\n`);
-
-    if (config.pgStart) {
-      stdout('Starting sweep PostgreSQL...\n');
-      await startPostgres();
-      postgresStarted = true;
-    }
-
-    for (const point of points) {
-      for (const rate of writeRates) {
-        stdout(`\n--> Running ${pointLabel(point)} @ ${rate} writes/s...\n`);
-        const attempt = await runAttempt({
-          config,
-          point,
-          writeRate: rate,
-          repetition: 0,
-          runsDir,
-          childLogsDir,
-        });
-        attempts.push(attempt);
-        await appendFile(attemptsPath, `${JSON.stringify(attempt)}\n`);
-
-        const s = attempt.summary;
-        const status = attempt.status === 'pass' ? 'PASS' : 'FAIL';
-        stdout(
-          `    Result: ${status} | achieved=${s?.achievedWriteRate.toFixed(1) ?? '0'} w/s | p95Lag=${s?.p95ClientVisibleLagMs.toFixed(1) ?? 'N/A'}ms\n`,
-        );
-
-        if (attempt.status === 'error' && !config.continueOnError) {
-          throw new Error(
-            `Benchmark attempt failed for ${pointLabel(point)} at ${rate} writes/s. See ${attempt.logPath}`,
-          );
-        }
-      }
-    }
-  } finally {
-    process.off('SIGINT', onSigint);
-    if (postgresStarted) {
-      stdout('\nStopping sweep PostgreSQL...\n');
-      await stopPostgres();
-    }
-  }
-
-  printLinearSweepSummaryTable(attempts);
-}
-
-function printLinearSweepSummaryTable(attempts: readonly SweepAttempt[]): void {
-  stdout(
-    '\n========================================================================================================================\n',
-  );
-  stdout(
-    '                                LINEAR THROUGHPUT SWEEP RESULTS                                                         \n',
-  );
-  stdout(
-    '========================================================================================================================\n',
-  );
-  stdout(
-    ' Profile     | Model     | Users | Rate  | Actual | E2E Lag (p50/p95) | IVM Adv (p50/p95) | RM Lag (p50/p95) | Status   \n',
-  );
-  stdout(
-    '-------------+-----------+-------+-------+--------+-------------------+-------------------+------------------+----------\n',
-  );
-
-  for (const a of attempts) {
-    const s = a.summary;
-    const p = a.point;
-    const e2eStr = s?.e2eServingLagMs
-      ? `${s.e2eServingLagMs.p50.toFixed(1)} / ${s.e2eServingLagMs.p95.toFixed(1)}ms`
-      : s?.p95ClientVisibleLagMs !== undefined
-        ? `${s.p50ClientVisibleLagMs.toFixed(1)} / ${s.p95ClientVisibleLagMs.toFixed(1)}ms`
-        : 'N/A';
-    const ivmStr = s?.advancementLatencyMs
-      ? `${s.advancementLatencyMs.p50.toFixed(1)} / ${s.advancementLatencyMs.p95.toFixed(1)}ms`
-      : 'N/A';
-    const rmStr = s?.replicationLagMs
-      ? `${s.replicationLagMs.p50.toFixed(1)} / ${s.replicationLagMs.p95.toFixed(1)}ms`
-      : 'N/A';
-    const status = a.status === 'pass' ? 'HEALTHY' : 'COLLAPSED';
-
-    stdout(
-      ` ${p.profile.padEnd(11)} | ` +
-        `${p.model.padEnd(9)} | ` +
-        `${String(p.users).padEnd(5)} | ` +
-        `${String(a.writeRate).padEnd(5)} | ` +
-        `${String(s?.achievedWriteRate.toFixed(1) ?? '0').padEnd(6)} | ` +
-        `${e2eStr.padEnd(17)} | ` +
-        `${ivmStr.padEnd(17)} | ` +
-        `${rmStr.padEnd(16)} | ` +
-        `${status}\n`,
-    );
-  }
-  stdout(
-    '========================================================================================================================\n\n',
-  );
-}
-
-async function runPointSearch(args: {
-  readonly config: SweepConfig;
-  readonly point: SweepPoint;
-  readonly runsDir: string;
-  readonly childLogsDir: string;
-  readonly attemptsPath: string;
-}): Promise<PointResult> {
-  let low = args.config.writeRateMin;
-  let high = args.config.writeRateMax;
-  let bestWriteRate: number | undefined;
-  let bestAttempt: SweepAttempt | undefined;
-  const attempts: SweepAttempt[] = [];
-
-  for (let step = 0; step < args.config.searchSteps && low <= high; step++) {
-    const writeRate = Math.floor((low + high) / 2);
-    const rateAttempts: SweepAttempt[] = [];
-    let ratePassed = true;
-
-    stdout(`  step ${step + 1}: ${writeRate} logical writes/s\n`);
-    for (
-      let repetition = 0;
-      repetition < args.config.repetitions;
-      repetition++
-    ) {
-      const attempt = await runAttempt({
-        config: args.config,
-        point: args.point,
-        writeRate,
-        repetition,
-        runsDir: args.runsDir,
-        childLogsDir: args.childLogsDir,
-      });
-      attempts.push(attempt);
-      rateAttempts.push(attempt);
-      await appendFile(args.attemptsPath, `${JSON.stringify(attempt)}\n`);
-
-      if (attempt.status === 'error' && !args.config.continueOnError) {
-        throw new Error(
-          `Benchmark attempt failed for ${pointLabel(args.point)} at ${writeRate} writes/s. See ${attempt.logPath}`,
-        );
-      }
-      if (attempt.status !== 'pass') {
-        ratePassed = false;
-        break;
-      }
-    }
-
-    if (ratePassed) {
-      bestWriteRate = writeRate;
-      bestAttempt = rateAttempts.at(-1);
-      low = writeRate + 1;
-    } else {
-      high = writeRate - 1;
-    }
-  }
-
-  return {
-    point: args.point,
-    bestWriteRate,
-    lowerBoundWriteRate: low,
-    upperBoundWriteRate: high,
-    attempts,
-    bestAttempt,
-  };
-}
-
-async function runAttempt(args: {
-  readonly config: SweepConfig;
-  readonly point: SweepPoint;
-  readonly writeRate: number;
-  readonly repetition: number;
-  readonly runsDir: string;
-  readonly childLogsDir: string;
-}): Promise<SweepAttempt> {
-  const outputPath = join(
-    args.runsDir,
-    `${pointID(args.point)}-rate${args.writeRate}-rep${args.repetition + 1}.json`,
-  );
-  const logPath = join(
-    args.childLogsDir,
-    `${pointID(args.point)}-rate${args.writeRate}-rep${args.repetition + 1}.log`,
-  );
-
-  if (args.config.resume && (await fileExists(outputPath))) {
-    const result = await readBenchmarkResult(outputPath);
-    return attemptFromResult({
-      point: args.point,
-      writeRate: args.writeRate,
-      repetition: args.repetition,
-      outputPath,
-      logPath,
-      reused: true,
-      result,
-    });
-  }
-
-  const command = benchmarkCommand(args.config, args.point, args.writeRate, {
-    outputPath,
-    logsDir: join(args.config.outputDir, 'logs'),
-  });
-  const exitCode = await runCommandToLog({
-    command: command[0],
-    args: command.slice(1),
-    cwd: repoRoot,
-    logPath,
-    verbose: args.config.verboseChildLogs,
-  });
-
-  if (!(await fileExists(outputPath))) {
-    return {
-      point: args.point,
-      writeRate: args.writeRate,
-      repetition: args.repetition,
-      outputPath,
-      logPath,
-      reused: false,
-      status: 'error',
-      exitCode,
-      error: `Benchmark exited ${exitCode} without writing ${outputPath}`,
-      summary: undefined,
-    };
-  }
-
-  const result = await readBenchmarkResult(outputPath);
-  return attemptFromResult({
-    point: args.point,
-    writeRate: args.writeRate,
-    repetition: args.repetition,
-    outputPath,
-    logPath,
-    reused: false,
-    result,
-    exitCode,
-  });
-}
-
-function attemptFromResult(args: {
-  readonly point: SweepPoint;
-  readonly writeRate: number;
-  readonly repetition: number;
-  readonly outputPath: string;
-  readonly logPath: string;
-  readonly reused: boolean;
-  readonly result: BenchmarkResult;
-  readonly exitCode?: number | undefined;
-}): SweepAttempt {
-  const pass = args.result.summary.pass;
-  return {
-    point: args.point,
-    writeRate: args.writeRate,
-    repetition: args.repetition,
-    outputPath: args.outputPath,
-    logPath: args.logPath,
-    reused: args.reused,
-    status: pass ? 'pass' : 'fail',
-    exitCode: args.exitCode,
-    error: undefined,
-    summary: args.result.summary,
-  };
-}
-
-function benchmarkCommand(
-  config: SweepConfig,
-  point: SweepPoint,
-  writeRate: number,
-  paths: {readonly outputPath: string; readonly logsDir: string},
-): readonly string[] {
-  const main = fileURLToPath(new URL('main.ts', import.meta.url));
-  const command = [
-    process.execPath,
-    main,
-    '--profile',
-    point.profile,
-    '--model',
-    point.model,
-    '--users',
-    String(point.users),
-    '--queries-per-user',
-    String(point.queriesPerUser),
-    '--rows-per-query',
-    String(point.rowsPerQuery),
-    '--write-rate',
-    String(writeRate),
-    '--duration-ms',
-    String(config.durationMs),
-    '--warmup-ms',
-    String(config.warmupMs),
-    '--settle-ms',
-    String(config.settleMs),
-    '--sample-interval-ms',
-    String(config.sampleIntervalMs),
-    '--progress-interval-ms',
-    String(config.progressIntervalMs),
-    '--slo-p99lag-ms',
-    String(config.sloP99LagMs),
-    '--batch-size',
-    String(config.batchSize),
-    '--payload-bytes',
-    String(config.payloadBytes),
-    '--zero-num-sync-workers',
-    String(point.zeroNumSyncWorkers),
-    '--zero-port',
-    String(config.zeroPort),
-    '--pg-start',
-    'false',
-    '--pg-stop-after-run',
-    'false',
-    '--output',
-    paths.outputPath,
-    '--logs-dir',
-    paths.logsDir,
-  ];
-  if (config.pgURL !== undefined) {
-    command.push('--pg-url', config.pgURL);
-  }
-  if (config.topology !== undefined) {
-    command.push('--topology', config.topology);
-  }
-  if (config.numViewSyncers !== undefined) {
-    command.push('--num-view-syncers', String(config.numViewSyncers));
-  }
-  if (config.numSyncWorkers !== undefined) {
-    command.push('--num-sync-workers', String(config.numSyncWorkers));
-  }
-  if (config.highAvailabilityRM) {
-    command.push('--high-availability-rm');
-  } else if (config.numReplicationManagers !== undefined) {
-    command.push(
-      '--num-replication-managers',
-      String(config.numReplicationManagers),
-    );
-  }
-  if (config.writeConcurrency !== undefined) {
-    command.push('--write-concurrency', String(config.writeConcurrency));
-  }
-  if (config.profileRM) {
-    command.push('--profile-rm');
-  }
-  if (config.profileVS) {
-    command.push('--profile-vs');
-  }
-  return command;
-}
-
-async function runCommandToLog(args: {
-  readonly command: string;
-  readonly args: readonly string[];
-  readonly cwd: string;
-  readonly logPath: string;
-  readonly verbose: boolean;
-}): Promise<number> {
-  await mkdir(dirname(args.logPath), {recursive: true});
-  const log = createWriteStream(args.logPath, {flags: 'w'});
-  await writeLog(log, `$ ${args.command} ${args.args.join(' ')}\n`);
-  try {
-    return await new Promise<number>((resolve, reject) => {
-      const child = spawn(args.command, args.args, {
-        cwd: args.cwd,
-        stdio: ['ignore', 'pipe', 'pipe'],
-      });
-      child.stdout?.on('data', chunk => {
-        log.write(chunk);
-        if (args.verbose) {
-          process.stdout.write(chunk);
-        }
-      });
-      child.stderr?.on('data', chunk => {
-        log.write(chunk);
-        if (args.verbose) {
-          process.stderr.write(chunk);
-        }
-      });
-      child.once('error', reject);
-      child.once('exit', code => {
-        resolve(code ?? 1);
-      });
-    });
-  } finally {
-    await closeLog(log);
-  }
-}
-
-function sweepPoints(config: SweepConfig): readonly SweepPoint[] {
-  const points: SweepPoint[] = [];
-  for (const profile of config.profiles) {
-    for (const model of config.models) {
-      for (const users of config.users) {
-        for (const rowsPerQuery of config.rowsPerQuery) {
-          for (const zeroNumSyncWorkers of config.syncWorkers) {
-            points.push({
-              profile,
-              model,
-              users,
-              queriesPerUser: config.queriesPerUser,
-              rowsPerQuery,
-              zeroNumSyncWorkers,
-            });
-            if (config.limit !== undefined && points.length >= config.limit) {
-              return points;
-            }
-          }
-        }
-      }
-    }
-  }
-  return points;
-}
-
-function parseArgs(argv: readonly string[]): SweepConfig {
-  const runID = new Date().toISOString().replace(/[:.]/g, '-');
-  let mode: 'binary' | 'linear' = 'binary';
-  let writeRates: readonly number[] | undefined;
-  let topology: BenchmarkTopology | undefined;
-  let numViewSyncers: number | undefined;
-  let numSyncWorkers: number | undefined;
-  let highAvailabilityRM: boolean | undefined;
-  let numReplicationManagers: (1 | 2) | undefined;
-  let writeConcurrency: number | undefined;
-  let profileRM = false;
-  let profileVS = false;
-  let profiles: readonly BenchmarkProfile[] = DEFAULT_PROFILES;
-  let models: readonly BenchmarkModel[] = DEFAULT_MODELS;
-  let users: readonly number[] = DEFAULT_USERS;
-  let rowsPerQuery: readonly number[] = DEFAULT_ROWS_PER_QUERY;
-  let syncWorkers: readonly number[] = DEFAULT_SYNC_WORKERS;
-  let queriesPerUser = 3;
-  let durationMs = 300_000;
-  let warmupMs = 30_000;
-  let settleMs = 5_000;
-  let sampleIntervalMs = 1_000;
-  let progressIntervalMs = 5_000;
-  let sloP99LagMs = 2_000;
-  let batchSize = 1;
-  let payloadBytes = 256;
-  let writeRateMin = 1;
-  let writeRateMax = DEFAULT_WRITE_RATE_MAX;
-  let searchSteps = DEFAULT_SEARCH_STEPS;
-  let repetitions = 1;
-  let outputDir = join('results', 'sweeps', runID);
-  let zeroPort = 4_848;
-  let pgStart = true;
-  let pgURL: string | undefined;
-  let resume = true;
-  let continueOnError = false;
-  let dryRun = false;
-  let limit: number | undefined;
-  let verboseChildLogs = false;
-  let help = false;
-
-  for (let i = 0; i < argv.length; i++) {
-    if (argv[i] === '--') {
-      continue;
-    }
-    const option = parseOption(argv[i]);
-    switch (option.name) {
-      case '--help':
-      case '-h':
-        help = true;
-        break;
-      case '--profiles':
-      case '--profile':
-        profiles = parseProfiles(readOptionValue(argv, option, i));
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--models':
-      case '--model':
-        models = parseModels(readOptionValue(argv, option, i));
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--users':
-      case '--user':
-        users = parsePositiveIntegerList(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--rows-per-query':
-        rowsPerQuery = parsePositiveIntegerList(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--sync-workers':
-      case '--zero-num-sync-workers':
-        syncWorkers = parsePositiveIntegerList(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--queries-per-user':
-        queriesPerUser = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--duration-ms':
-        durationMs = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--warmup-ms':
-        warmupMs = parseNonNegativeInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--settle-ms':
-        settleMs = parseNonNegativeInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--sample-interval-ms':
-        sampleIntervalMs = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--progress-interval-ms':
-        progressIntervalMs = parseNonNegativeInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--slo-p99-lag-ms':
-      case '--slo-p99lag-ms':
-        sloP99LagMs = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--batch-size':
-        batchSize = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--payload-bytes':
-        payloadBytes = parseNonNegativeInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--write-rate-min':
-        writeRateMin = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--write-rate-max':
-        writeRateMax = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--search-steps':
-        searchSteps = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--repetitions':
-        repetitions = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--output-dir':
-        outputDir = readOptionValue(argv, option, i);
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--zero-port':
-        zeroPort = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--pg-start': {
-        const parsed = readBooleanOption(argv, option, i, true);
-        pgStart = parsed.value;
-        i += parsed.consumed;
-        break;
-      }
-      case '--pg-url':
-        pgURL = readOptionValue(argv, option, i);
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--resume': {
-        const parsed = readBooleanOption(argv, option, i, true);
-        resume = parsed.value;
-        i += parsed.consumed;
-        break;
-      }
-      case '--continue-on-error': {
-        const parsed = readBooleanOption(argv, option, i, true);
-        continueOnError = parsed.value;
-        i += parsed.consumed;
-        break;
-      }
-      case '--dry-run': {
-        const parsed = readBooleanOption(argv, option, i, true);
-        dryRun = parsed.value;
-        i += parsed.consumed;
-        break;
-      }
-      case '--limit':
-        limit = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      case '--verbose-child-logs': {
-        const parsed = readBooleanOption(argv, option, i, true);
-        verboseChildLogs = parsed.value;
-        i += parsed.consumed;
-        break;
-      }
-      case '--mode': {
-        const val = readOptionValue(argv, option, i);
-        if (val !== 'binary' && val !== 'linear') {
-          throw new Error('--mode must be binary or linear');
-        }
-        mode = val;
-        i += option.value === undefined ? 1 : 0;
-        break;
-      }
-      case '--write-rates': {
-        writeRates = parsePositiveIntegerList(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        mode = 'linear';
-        i += option.value === undefined ? 1 : 0;
-        break;
-      }
-      case '--topology': {
-        const val = readOptionValue(argv, option, i);
-        if (val !== 'single' && val !== 'distributed') {
-          throw new Error('--topology must be single or distributed');
-        }
-        topology = val;
-        i += option.value === undefined ? 1 : 0;
-        break;
-      }
-      case '--num-view-syncers': {
-        numViewSyncers = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      }
-      case '--high-availability-rm': {
-        const parsed = readBooleanOption(argv, option, i, true);
-        highAvailabilityRM = parsed.value;
-        i += parsed.consumed;
-        break;
-      }
-      case '--num-sync-workers': {
-        numSyncWorkers = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      }
-      case '--num-replication-managers': {
-        const val = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        if (val !== 1 && val !== 2) {
-          throw new Error('--num-replication-managers must be 1 or 2');
-        }
-        numReplicationManagers = val;
-        i += option.value === undefined ? 1 : 0;
-        break;
-      }
-      case '--write-concurrency': {
-        writeConcurrency = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      }
-      case '--rows-per-tx': {
-        batchSize = parsePositiveInteger(
-          option.name,
-          readOptionValue(argv, option, i),
-        );
-        i += option.value === undefined ? 1 : 0;
-        break;
-      }
-      case '--profile-rm': {
-        const parsed = readBooleanOption(argv, option, i, true);
-        profileRM = parsed.value;
-        i += parsed.consumed;
-        break;
-      }
-      case '--profile-vs': {
-        const parsed = readBooleanOption(argv, option, i, true);
-        profileVS = parsed.value;
-        i += parsed.consumed;
-        break;
-      }
-      default:
-        throw new Error(`Unknown sweep option ${option.name}`);
-    }
-  }
-
-  if (help) {
-    printUsage();
-    process.exit(0);
-  }
-  if (writeRateMin > writeRateMax) {
-    throw new Error('--write-rate-min must be <= --write-rate-max');
-  }
-
-  return {
-    runID,
-    mode,
-    writeRates,
-    topology,
-    numViewSyncers,
-    numSyncWorkers,
-    highAvailabilityRM,
-    numReplicationManagers,
-    writeConcurrency,
-    profileRM,
-    profileVS,
-    profiles,
-    models,
-    users,
-    rowsPerQuery,
-    syncWorkers,
-    queriesPerUser,
-    durationMs,
-    warmupMs,
-    settleMs,
-    sampleIntervalMs,
-    progressIntervalMs,
-    sloP99LagMs,
-    batchSize,
-    payloadBytes,
-    writeRateMin,
-    writeRateMax,
-    searchSteps,
-    repetitions,
-    outputDir,
-    zeroPort,
-    pgStart,
-    pgURL,
-    resume,
-    continueOnError,
-    dryRun,
-    limit,
-    verboseChildLogs,
-  };
-}
-
-function parseOption(arg: string): {
-  readonly name: string;
-  readonly value: string | undefined;
-} {
-  const equals = arg.indexOf('=');
-  if (equals === -1) {
-    return {name: arg, value: undefined};
-  }
-  return {name: arg.slice(0, equals), value: arg.slice(equals + 1)};
-}
-
-function readOptionValue(
-  argv: readonly string[],
-  option: {readonly name: string; readonly value: string | undefined},
-  index: number,
-): string {
-  if (option.value !== undefined) {
-    return option.value;
-  }
-  const value = argv[index + 1];
-  if (value === undefined || value.startsWith('--')) {
-    throw new Error(`${option.name} requires a value`);
-  }
-  return value;
-}
-
-function readBooleanOption(
-  argv: readonly string[],
-  option: {readonly name: string; readonly value: string | undefined},
-  index: number,
-  defaultValue: boolean,
-): {readonly value: boolean; readonly consumed: number} {
-  if (option.value !== undefined) {
-    return {value: parseBoolean(option.value), consumed: 0};
-  }
-  const value = argv[index + 1];
-  if (value === 'true' || value === 'false') {
-    return {value: parseBoolean(value), consumed: 1};
-  }
-  return {value: defaultValue, consumed: 0};
-}
-
-function parseProfiles(value: string): readonly BenchmarkProfile[] {
-  const profiles = value.split(',').map(part => {
-    const trimmed = part.trim();
-    if (!PROFILE_VALUES.has(trimmed as BenchmarkProfile)) {
-      throw new Error(`Invalid profile "${trimmed}"`);
-    }
-    return trimmed as BenchmarkProfile;
-  });
-  if (profiles.length === 0) {
-    throw new Error('--profiles must not be empty');
-  }
-  return profiles;
-}
-
-function parseModels(value: string): readonly BenchmarkModel[] {
-  const models = value.split(',').map(part => {
-    const trimmed = part.trim();
-    if (!MODEL_VALUES.has(trimmed as BenchmarkModel)) {
-      throw new Error(`Invalid model "${trimmed}"`);
-    }
-    return trimmed as BenchmarkModel;
-  });
-  if (models.length === 0) {
-    throw new Error('--models must not be empty');
-  }
-  return models;
-}
-
-function parsePositiveIntegerList(
-  name: string,
-  value: string,
-): readonly number[] {
-  const values = value
-    .split(',')
-    .map(part => parsePositiveInteger(name, part.trim()));
-  if (values.length === 0) {
-    throw new Error(`${name} must not be empty`);
-  }
-  return values;
-}
-
-function parsePositiveInteger(name: string, value: string): number {
-  const parsed = Number(value);
-  if (!Number.isInteger(parsed) || parsed <= 0) {
-    throw new Error(`${name} must be a positive integer`);
-  }
-  return parsed;
-}
-
-function parseNonNegativeInteger(name: string, value: string): number {
-  const parsed = Number(value);
-  if (!Number.isInteger(parsed) || parsed < 0) {
-    throw new Error(`${name} must be a non-negative integer`);
-  }
-  return parsed;
-}
-
-function parseBoolean(value: string): boolean {
-  switch (value) {
-    case 'true':
-      return true;
-    case 'false':
+if (isLinear) {
+  const filtered = cleanArgv.filter((arg, i) => {
+    if (arg === '--mode=linear') {
       return false;
-    default:
-      throw new Error(`Expected boolean value, got "${value}"`);
-  }
-}
-
-function printDryRun(config: SweepConfig): void {
-  const points = sweepPoints(config);
-  if (config.mode === 'linear') {
-    const rates = config.writeRates ?? [100, 250, 500, 1000, 2000];
+    }
+    if (arg === '--mode') {
+      return false;
+    }
+    if (i > 0 && cleanArgv[i - 1] === '--mode') {
+      return false;
+    }
+    return true;
+  });
+  const {parseLinearSweepArgs, runLinearSweep} =
+    await import('./linear-sweep.ts');
+  const {sweepPoints, pointLabel, stdout} = await import('./sweep-common.ts');
+  const config = parseLinearSweepArgs(filtered);
+  if (config.dryRun) {
+    const points = sweepPoints(config);
     stdout(`zero-throughput linear sweep dry run\n`);
-    stdout(`points: ${points.length}, write rates: ${rates.join(', ')}\n`);
-    stdout(`total benchmark runs: ${points.length * rates.length}\n`);
-    stdout(`duration per benchmark: ${formatDuration(config.durationMs)}\n\n`);
+    stdout(
+      `points: ${points.length}, write rates: ${config.writeRates.join(', ')}\n`,
+    );
+    stdout(
+      `total benchmark runs: ${points.length * config.writeRates.length}\n\n`,
+    );
     for (const point of points) {
-      for (const rate of rates) {
+      for (const rate of config.writeRates) {
         stdout(`${pointLabel(point)} @ ${rate} writes/s\n`);
       }
     }
-    return;
+  } else {
+    await runLinearSweep(config);
   }
-  stdout(`zero-throughput recommended sweep dry run\n`);
-  stdout(`points: ${points.length}\n`);
-  stdout(
-    `max benchmark runs: ${points.length * config.searchSteps * config.repetitions}\n`,
-  );
-  stdout(`duration per benchmark: ${formatDuration(config.durationMs)}\n`);
-  stdout(
-    `write-rate search: ${config.writeRateMin}-${config.writeRateMax} logical writes/s, ${config.searchSteps} steps\n\n`,
-  );
-  for (const point of points) {
-    stdout(`${pointLabel(point)}\n`);
+} else {
+  const {parseBinarySweepArgs, runBinarySweep} =
+    await import('./binary-sweep.ts');
+  const {sweepPoints, pointLabel, stdout} = await import('./sweep-common.ts');
+  const config = parseBinarySweepArgs(cleanArgv);
+  if (config.dryRun) {
+    const points = sweepPoints(config);
+    stdout(`zero-throughput binary search sweep dry run\n`);
+    stdout(`points: ${points.length}\n`);
+    stdout(
+      `max benchmark runs: ${points.length * config.searchSteps * config.repetitions}\n`,
+    );
+    stdout(
+      `write-rate search: ${config.writeRateMin}-${config.writeRateMax} logical writes/s, ${config.searchSteps} steps\n\n`,
+    );
+    for (const point of points) {
+      stdout(`${pointLabel(point)}\n`);
+    }
+  } else {
+    await runBinarySweep(config);
   }
-}
-
-function printUsage(): void {
-  stdout(`Usage:
-  pnpm --filter zero-throughput run sweep -- [options]
-
-Default matrix:
-  --profiles relational,email,forum
-  --models hot
-  --users 50,100,200,400
-  --rows-per-query 50
-  --sync-workers 1,2,4
-  --queries-per-user 3
-  --duration-ms 300000
-
-Write-rate search:
-  --write-rate-min 1
-  --write-rate-max 100
-  --search-steps 7
-
-Useful:
-  --dry-run
-  --limit 1
-  --output-dir results/sweeps/my-run
-  --models hot,realistic
-  --pg-start false
-  --verbose-child-logs
-`);
-}
-
-async function readBenchmarkResult(path: string): Promise<BenchmarkResult> {
-  const parsed = JSON.parse(await readFile(path, 'utf8')) as unknown;
-  if (
-    parsed === null ||
-    typeof parsed !== 'object' ||
-    !('summary' in parsed) ||
-    parsed.summary === null ||
-    typeof parsed.summary !== 'object' ||
-    !('pass' in parsed.summary) ||
-    typeof parsed.summary.pass !== 'boolean'
-  ) {
-    throw new Error(`${path} is not a benchmark result`);
-  }
-  return parsed as BenchmarkResult;
-}
-
-async function fileExists(path: string): Promise<boolean> {
-  try {
-    await access(path);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-async function writeLog(stream: WriteStream, text: string): Promise<void> {
-  if (!stream.write(text)) {
-    await once(stream, 'drain');
-  }
-}
-
-async function closeLog(stream: WriteStream): Promise<void> {
-  stream.end();
-  await once(stream, 'finish');
-}
-
-function csvHeader(): string {
-  return [
-    'profile',
-    'model',
-    'users',
-    'queriesPerUser',
-    'rowsPerQuery',
-    'zeroNumSyncWorkers',
-    'bestWriteRate',
-    'attemptedWriteRates',
-    'bestP99ClientVisibleLagMs',
-    'bestMaxSeqLag',
-    'bestLagSlopeSeqPerSec',
-    'bestAffectedActiveClientGroupWriteRatio',
-    'bestOutputPath',
-    'bestFailureReasons',
-  ].join(',');
-}
-
-function csvRow(result: PointResult): string {
-  const best = result.bestAttempt;
-  return [
-    result.point.profile,
-    result.point.model,
-    result.point.users,
-    result.point.queriesPerUser,
-    result.point.rowsPerQuery,
-    result.point.zeroNumSyncWorkers,
-    result.bestWriteRate ?? '',
-    result.attempts.map(attempt => attempt.writeRate).join('|'),
-    best?.summary?.p99ClientVisibleLagMs ?? '',
-    best?.summary?.maxSeqLag ?? '',
-    best?.summary?.lagSlopeSeqPerSec ?? '',
-    best?.summary?.writeImpact.affectedActiveClientGroupWriteRatio ?? '',
-    best?.outputPath ?? '',
-    best?.summary?.failureReasons.join('|') ?? '',
-  ]
-    .map(csvCell)
-    .join(',');
-}
-
-function csvCell(value: unknown): string {
-  const text = String(value);
-  if (CSV_ESCAPE_PATTERN.test(text)) {
-    return `"${text.replaceAll('"', '""')}"`;
-  }
-  return text;
-}
-
-function pointID(point: SweepPoint): string {
-  return [
-    point.profile,
-    point.model,
-    `${point.users}u`,
-    `${point.queriesPerUser}q`,
-    `${point.rowsPerQuery}rows`,
-    `${point.zeroNumSyncWorkers}sync`,
-  ].join('-');
-}
-
-function pointLabel(point: SweepPoint): string {
-  return `${point.profile}:${point.model} users=${point.users} queriesPerUser=${point.queriesPerUser} rowsPerQuery=${point.rowsPerQuery} syncWorkers=${point.zeroNumSyncWorkers}`;
-}
-
-function gitCommit(): string | undefined {
-  try {
-    return execFileSync('git', ['rev-parse', 'HEAD'], {
-      cwd: appRoot,
-      encoding: 'utf8',
-      stdio: ['ignore', 'pipe', 'ignore'],
-    }).trim();
-  } catch {
-    return undefined;
-  }
-}
-
-function stdout(message: string): void {
-  process.stdout.write(message);
-}
-
-function stderr(message: string): void {
-  process.stderr.write(message);
 }
